backend version 1.0

+++ backend/Controllers/UserController.js
@@ -1,6 +1,9 @@
 import User from '../models/UserModel.js'
+import bcrypt from 'bcrypt'
+
 
 export const getUsers = async (req, res) => {
+    if (!req.session.userId) return res.status(401).json({ msg: 'Your login session has expired' })
     try {
         const response = await User.findAll();
         res.status(200).json(response);
@@ -11,6 +14,7 @@ export const getUsers = async (req, res) => {
 }
 
 export const getUserById = async (req, res) => {
+    if (!req.session.userId) return res.status(401).json({ msg: 'Your login session has expired' })
     try {
         const response = await User.findOne({
             where: {
@@ -25,8 +29,23 @@ export const getUserById = async (req, res) => {
 }
 
 export const createUser = async (req, res) => {
+
+    const { fullName, userName, userEmail, userPhone, userDob, userGender, userAddress, userPassword, userConfirmPassword, userTerms } = req.body;
+    if (userPassword !== userConfirmPassword) return res.status(400).json({ msg: 'Password and Confirm Password do not match' });
+    const salt = await bcrypt.genSalt();
+    const hashPassword = await bcrypt.hash(userPassword, salt);
     try {
-        await User.create(req.body);
+        await User.create({
+            fullname: fullName,
+            username: userName,
+            email: userEmail,
+            phone: userPhone,
+            Dob: userDob,
+            Gender: userGender,
+            address: userAddress,
+            password: hashPassword,
+            terms: userTerms
+        });
         res.status(201).json({ msg: "User Created" });
     } catch (error) {
         error.errors.map(e => {
@@ -37,10 +56,35 @@ export const createUser = async (req, res) => {
 }
 
 export const updateUser = async (req, res) => {
+    if (!req.session.userId) return res.status(401).json({msg: 'Your login session has expired'})
+    const { fullName, userName, userEmail, userPhone, userDob, userGender, userAddress, useroldPassword, userPassword, userConfirmPassword } = req.body;
+
+    const user = await User.findOne({
+        where: {
+            uid: req.params.id
+        }
+    })
+
+    if (!user) return res.status(404).json({ msg: 'User id not found' })
+    if (!useroldPassword !== user.password) return res.status(404).json({ msg: 'Old password does not matched' })
+    if (useroldPassword === "" || useroldPassword === null || userPassword === "" || userPassword === null || userConfirmPassword === "" || userConfirmPassword === null) return res.status(404).json({ msg: 'Blank fields not allowed' })
+    if (userPassword !== userConfirmPassword) return res.status(400).json({ msg: 'Password and Confirm Password do not match' });
+    const salt = await bcrypt.genSalt();
+    const NewPassword = await bcrypt.hash(userPassword, salt);
+
     try {
-        await User.update(req.body, {
+        await User.update({
+            fullname: fullName,
+            username: userName,
+            email: userEmail,
+            phone: userPhone,
+            Dob: userDob,
+            Gender: userGender,
+            address: userAddress,
+            password: NewPassword,
+        }, {
             where: {
-                uid: req.params.id
+                uid: user.id
             }
         });
         res.status(200).json({ msg: "User Updated" });
@@ -51,6 +95,7 @@ export const updateUser = async (req, res) => {
 }
 
 export const deleteUser = async (req, res) => {
+    if (!req.session.userId) return res.status(401).json({msg: 'Your login session has expired'})
     try {
         await User.destroy({
             where: {
